requirements.txt
google-cloud-storage==1.7.0